change stock type from decimal to long

+++ Tibr.Application/Dtos/ProductDto/CreateProductDto.cs
@@ -14,7 +14,7 @@ namespace Tibr.Application.Dtos.ProductDto
         public decimal Weight { get; set; }
         public decimal BuyPrice { get; set; }
         public decimal SellPrice { get; set; }
-        public decimal Stock { get; set; }
+        public long Stock { get; set; }
         public string? ImageUrl { get; set; }
     }
 }
